feat: 보컬 분석 초기 페이지 구현 및 기존 페이지 경로 수정
